User creation moduel created, and new auth object made.
There is now a simple callable function in shiny auth that allows
you to add a new user to the db, spesifcying and id, a password, and
weather or not the user is an admin.

To make the curation of this module simpliar, and make calling
SBA modules simpliar for the user, auth_server now passes an extra
argument into the server funciton, the auth object, like the session
object for shiny it contains all of the information about the db
needed by the auth modules.  This for example means that to call the
create new user function you only need provide one argument, auth.