Settings is now a part of the ShinyBasicAuth pacakge
The Settings tab from TalentFramework has been moved into ShinyBascAuth.
It is created off of the auth_config file that creates the users table
in the db.

There has also been changes to the auth object, the way that it is
created and what it contains.  It now contains the list form auth_config
so that the settings tab can be created.  Auth also contains dt_user to
simplify app creation.