Bug fixes
There were two bugs that have been fixed.  Before any columns added
to the auth_conf as categorical would not be added to the db and no
error would be given, this has been fixed and now if an unregonised
type is given there is a error.

The other error was that the inital user would have an AN for
modorator if it was being used and this would resut in a error.