db should now be split across two scema
To keep the users and the data seperate there are now two pool objects
in auth, pool_auth and pool_data.  They keep the apps data and the auth
information seperate. In the fureute this may allow the app to be more
sercure, and for the time being makes the db more managable.

Also the use of pool inset of regular conns is now inforced.