Added test of given server function to auth_server
Now auth_server will test the function given to it when the server
is started.  It will stop with an error if the arguments are not
correct.

Also some formatting and ui chages to sidebar that make the code
more readable and change the font size of the text input titles.