Fix to Changeing the password from Admin
There was a very servious problem that cased changes to a password
from the admin tab to defult to changin the password of the loged
on user, not the user that had been selected.

This has now been fixed and the change passowrd function will apply
basic tests to prevent this from happaning again.