Moderatior is not shown in settings when not in use
There was a bug that caused an the settengs page to always try
to show the moderator box on the settings page, even when it was
not being used in auth.  This has been fixed so that it is not
shown if the use modorators flag is set to false.

There has also been a change to the boxes that are shown in the
settings tab, not the date created and last password change will
not be shown to the user.